+++ src/main/java/it/italiangrid/portal/dirac/db/dao/generic/JobsDAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import it.italiangrid.portal.dirac.db.domain.Jobs;
 
-public interface JobsDAO extends GenericDAO<Jobs, Integer>{
+public interface JobsDAO extends GenericDAO<Jobs, Long>{
 
 	List<Jobs> findByOwner(String owner);
